+++ src/lib/util/handlers.ts
@@ -7,3 +7,9 @@ export const onEnter = (fn: WantsKbdEv | Action) => (ev: KeyboardEvent) => {
     fn(ev);
   }
 }
+
+export const onTab = (fn: WantsKbdEv | Action) => (ev: KeyboardEvent) => {
+  if (ev.key === 'Tab') {
+    fn(ev);
+  }
+}