+++ src/routes/character/[id]/DiceDialog.svelte
@@ -5,11 +5,13 @@
 
   let title = '';
   let dice: DieValue[] = [];
+  let msg = '';
 
-  export async function show(label: string, rolls: DieValue[] = []) {
+  export async function show(label: string, rolls: DieValue[] = [], subtext = '') {
     title = label;
     dice = rolls;
+    msg = subtext;
     return dialog.open();
   }
 </script>
-<NotificationDialog {title} {dice} bind:this={dialog}/>
+<NotificationDialog {title} {dice} {msg} bind:this={dialog}/>